homework/requirements.txt
diskcache==5.6.3
Jinja2==3.1.6
llama_cpp_python==0.3.20
MarkupSafe==3.0.3
numpy==2.4.4
typing_extensions==4.15.0
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
asyncpg>=0.30.0
alembic>=1.13.1
redis==5.0.4
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.0
python-dotenv==1.0.1
